bcc: Generate tracepoint argument structures automatically
Modify the BPF module to generate tracepoint argument structures
automatically when encountering their declarations as a function
argument. This is a work-in-progress, for example there's no
support for loading from source file yet (only from string). Also,
this is a fragile approach -- it would be best to perform this on
the clang side and not on the Python side.